moved renderAllocations into their own component, updated SiteTable to call that component
